requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
mcp>=1.9.0
boto3>=1.34.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0
httpx>=0.27.0
langgraph>=0.4.0
langchain-aws>=0.2.0
langchain-mcp-adapters>=0.1.0
